chore: align playground deps
