+++ playgrounds/next/package.json
@@ -1,31 +1,25 @@
 {
   "name": "next-app",
   "private": true,
+  "type": "module",
   "scripts": {
-    "analyze": "ANALYZE=true next build",
-    "dev": "next dev",
-    "build": "next build",
+    "dev": "next dev --webpack",
+    "build": "next build --webpack",
     "start": "next start",
     "lint": "next lint"
   },
   "dependencies": {
     "@next/bundle-analyzer": "^14.2.3",
-    "@tanstack/react-query": ">=5.45.1",
-    "next": "^15.5.4",
-    "react": ">=18.3.1",
-    "react-dom": ">=18.3.1",
+    "@tanstack/react-query": "catalog:",
+    "next": "^16.0.3",
+    "react": "catalog:",
+    "react-dom": "catalog:",
     "viem": "2.*",
     "wagmi": "workspace:*"
   },
   "devDependencies": {
     "@types/node": "^24.5.1",
-    "@types/react": ">=18.3.1",
-    "@types/react-dom": ">=18.3.0",
-    "bufferutil": "^4.0.8",
-    "encoding": "^0.1.13",
-    "lokijs": "^1.5.12",
-    "pino-pretty": "^10.3.1",
-    "supports-color": "^9.4.0",
-    "utf-8-validate": "^5.0.2"
+    "@types/react": "catalog:",
+    "@types/react-dom": "catalog:"
   }
 }
