+++ playgrounds/nuxt/package.json
@@ -10,11 +10,11 @@
     "_postinstall": "nuxt prepare"
   },
   "dependencies": {
-    "@tanstack/vue-query": ">=5.45.0",
+    "@tanstack/vue-query": "catalog:",
     "@wagmi/vue": "workspace:*",
-    "nuxt": "^3.19.2",
+    "nuxt": "catalog:",
     "viem": "2.*",
-    "vue": ">=3.4.21",
+    "vue": "catalog:",
     "vue-router": "^4.3.2"
   }
 }
