+++ playgrounds/vite-vue/package.json
@@ -8,14 +8,13 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@tanstack/vue-query": ">=5.45.0",
+    "@tanstack/vue-query": "catalog:",
     "@wagmi/vue": "workspace:*",
     "viem": "2.*",
-    "vue": ">=3.4.21"
+    "vue": "catalog:"
   },
   "devDependencies": {
     "@vitejs/plugin-vue": "catalog:",
-    "buffer": "^6.0.3",
     "vite": "catalog:",
     "vue-tsc": "3.1.3"
   }
