+++ pnpm-workspace.yaml
@@ -13,6 +13,7 @@ catalog:
   '@coinbase/wallet-sdk': 4.3.6
   '@gemini-wallet/core': 0.3.2
   '@metamask/sdk': 0.33.1
+  '@nuxt/schema': 4.2.1
   '@safe-global/safe-apps-provider': 0.18.6
   '@safe-global/safe-apps-sdk': 9.1.0
   '@tanstack/query-core': 5.49.1
@@ -23,6 +24,7 @@ catalog:
   '@vitejs/plugin-react': ^4.3.3
   '@vitejs/plugin-vue': 6.0.1
   '@walletconnect/ethereum-provider': 2.21.1
+  nuxt: 4.2.1
   porto: 0.2.35
   react: 19.2.0
   react-dom: 19.2.0
